02-use-cases/AWS-operations-agent/agent-lambda/src/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.11.5
boto3==1.39.12
botocore==1.39.12
aiohttp==3.12.14
python-multipart>=0.0.9
python-dotenv==1.0.0
# Strands SDK dependencies
strands-agents==0.1.6
mcp==1.10.0
pyjwt==2.10.1
cryptography==45.0.4
